requirements.txt
Flask==2.3.2
Flask-Cors==4.0.0
mysql-connector-python==8.3.0
requests==2.31.0
Werkzeug==2.3.6

